release: 0.1.2 — 网关一致性修复 + --version 同步 + 注释清理
握手
- EC-JPAKE 改为按 1 字节长度前缀读取变长 ZKP 标量与公钥点，
  并用游标 + 尾字节校验替换写死的长度断言（修复偶发握手失败）。

schema（遵循 MIoT-spec-v2）
- device-spec format 由枚举放宽为字符串；rule.version 改为可选。
- delay/loop 去掉 min(1)；device-get/input 的 siid/piid/aiid/eiid 放宽为整数。
- device-output 变量引用支持可选 min/max/step，并对 number dtype 强制范围校验。
- device-input-set-var 参数 scope 缺省时校验 id 为字符串。

校验 / lint
- validate-graph 增加 cfg.version 整数校验，拆分 delay（仅整数）与 statusLast（>0），
  接入 reachability 检查。
- lint-graph 增加 condition 入口未连线告警（网关实测缺省判 FALSE，仅未命中分支会触发）。

导出
- export-rule 将事件参数导出为 --event-filter，多值场景补告警，空 day 数组跳过。

backup
- 新增 --wait：轮询操作进度到 100% 再返回（progress_id 为 0 即本地缓存命中，立即完成）。

CLI
- --version / agentVersion 由 0.1.0 同步为 0.1.2。

注释
- 移除源码注释中的内部备忘录引用，改为中性表述。

+++ packages/core/src/schemas/nodes/device-get-set-var.ts
@@ -23,8 +23,8 @@ export type DeviceGetSetVarCfg = z.infer<typeof DeviceGetSetVarCfg>;
 export const DeviceGetSetVarProps = z
   .object({
     did: z.string(),
-    siid: z.number(),
-    piid: z.number(),
+    siid: z.number().int(),
+    piid: z.number().int(),
     dtype: z.enum(['number', 'boolean', 'string']),
     scope: z.string(),
     id: z.string(),
